requirements.txt
anthropic>=0.30.0
replicate>=0.25.0
elevenlabs>=1.3.0
requests>=2.31.0
moviepy>=1.0.3
openai-whisper>=20231117
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
numpy>=1.24.0
